Initial commit of new configuration file management and versioning system. Highlights:
Configuration file: Now creates or uses ~/.molflow/config to handle configuration. Currently only uses the local workflow paths in that file, but it's fleshed out with more options that may be used in the future. Uses this configuration for local workflow paths to discover workflows on the local system. Note that the default includes ./, so it will discover workflows located in directories off the current.

Versioning system: Workflows are now expected to be stored separately from this repository, including the one 'required' workflow - convert. If they are git repositories, we can understand versions from those locations as tags, and can select versions to run with the run -v option. Note that git list has been updated with a verbose option to show all versions available.

NOTES: There is a lot of dead code and commented out code here, and the tests have not yet been updated. All the current stuff has been tested with basic add_two workflows, but should hopefully scale to other situations.